letsencrypt_nginx should not depend on letsencrypt_apache.
